split into two crawlers
