+++ tests/bigquery/usage/test_extractor.py
@@ -26,7 +26,9 @@ async def test_extractor(test_root_dir):
     entries = load_entries(test_root_dir + "/bigquery/usage/data/sample_log.json")
 
     # @patch doesn't work for async func in py3.7: https://bugs.python.org/issue36996
-    with patch("metaphor.bigquery.usage.extractor.build_client") as mock_build_client:
+    with patch(
+        "metaphor.bigquery.usage.extractor.build_logging_client"
+    ) as mock_build_client:
         mock_build_client.return_value.project = "project1"
         mock_list_entries(mock_build_client, entries)
